+++ src/error.rs
@@ -61,13 +61,11 @@ pub trait StackedErrorDowncast: StackableErrorTrait + Sized {
         E: Display + Send + Sync + 'static;
 }
 
-/// NOTE: this type is only public because `impl Trait` in associated types is
-/// unstable, only `StackedErrorDowncast` methods are intended to be used on
-/// this.
-// The specific type that `Error` uses in its stack. NOTE the `error_kind_size`
-// should be updated whenever this is changed. pub type ErrorBox = Box<dyn
-// Display + Send + Sync + 'static>;
-pub struct ErrorItem {
+/// The specific type that `Error` uses in its stack, should only be needed for
+/// low level manipulation and convenience.
+#[must_use]
+pub struct StackedErrorItem {
+    // NOTE the `error_kind_size` should be updated whenever this is changed
     b: SmallBox<dyn StackableErrorTrait, smallbox::space::S4>,
     l: Option<&'static Location<'static>>,
 }
@@ -75,10 +73,10 @@ pub struct ErrorItem {
 #[cfg(target_pointer_width = "64")]
 #[test]
 fn error_kind_size() {
-    assert_eq!(core::mem::size_of::<ErrorItem>(), 56);
+    assert_eq!(core::mem::size_of::<StackedErrorItem>(), 56);
 }
 
-impl ErrorItem {
+impl StackedErrorItem {
     pub fn new<E: Display + Send + Sync + 'static>(
         e: E,
         l: Option<&'static Location<'static>>,
@@ -87,7 +85,7 @@ impl ErrorItem {
     }
 }
 
-impl Debug for ErrorItem {
+impl Debug for StackedErrorItem {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         f.write_fmt(format_args!("{}", self.get_err()))?;
         if let Some(location) = self.get_location() {
@@ -97,13 +95,13 @@ impl Debug for ErrorItem {
     }
 }
 
-impl Display for ErrorItem {
+impl Display for StackedErrorItem {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         core::fmt::Debug::fmt(self, f)
     }
 }
 
-impl StackedErrorDowncast for ErrorItem {
+impl StackedErrorDowncast for StackedErrorItem {
     fn get_err(&self) -> &(impl Display + Send + Sync + 'static) {
         &self.b
     }
@@ -134,10 +132,10 @@ impl StackedErrorDowncast for ErrorItem {
     }
 }
 
-/// An error struct intended for high level error propogation with programmable
+/// An error struct intended for high level error propagation with programmable
 /// backtraces
 ///
-/// For lower level error propogation, you should still use ordinary [Option]
+/// For lower level error propagation, you should still use ordinary [Option]
 /// and [Result] with domain-specific enums, it is only when using OS-level
 /// functions or when multiple domains converge that this is intended to be
 /// used. This has an internal stack for different kinds of arbitrary errors and
@@ -145,15 +143,35 @@ impl StackedErrorDowncast for ErrorItem {
 /// [StackableErr](crate::StackableErr) trait, this enables easy conversion and
 /// software defined backtraces for better `async` debugging. See the crate docs
 /// for more.
+///
+/// Note: in most cases the stacking operations do the expected thing of pushing
+/// the element as-is onto the internal stack. However, the inherent addition
+/// methods on [StackedError] and the fundamental [crate::StackableErr] methods
+/// special-case [StackedError] itself by attempting to downcast it. If that
+/// succeeds, they perform the equivalent of [StackedError::chain_errors] to
+/// combine the two into a single [StackedError]. This prevents undesirable
+/// nesting and greatly improves the display when [StackedError]s combine.
+#[must_use]
 pub struct StackedError {
     /// Using a ThinVec has advantages such as taking as little space as
     /// possible on the stack (since we are commiting to some indirection at
     /// this point), and having the niche optimizations applied to things like
     /// `Result<(), Error>`.
-    stack: ThinVec<ErrorItem>,
+    stack: ThinVec<StackedErrorItem>,
 }
 
 pub type Error = StackedError;
+
+/// Sees if `e` is a `StackedError` and returns its stack without allocation
+fn take_error_stack<E: Display + Send + Sync + 'static>(
+    e: &mut E,
+) -> Option<ThinVec<StackedErrorItem>> {
+    let tmp: &mut dyn StackableErrorTrait = e;
+    // does not allocate
+    tmp._as_any_mut()
+        .downcast_mut::<Error>()
+        .map(|other| core::mem::take(&mut other.stack))
+}
 
 /// Note: in most cases you can use `Error::from` or a call from `StackableErr`
 /// instead of these functions.
@@ -172,15 +190,26 @@ impl Error {
     }
 
     #[track_caller]
-    pub fn from_err<E: Display + Send + Sync + 'static>(e: E) -> Self {
-        Self {
-            stack: thin_vec![ErrorItem::new(e, Some(Location::caller()))],
+    pub fn from_err<E: Display + Send + Sync + 'static>(mut e: E) -> Self {
+        if let Some(stack) = take_error_stack(&mut e) {
+            let mut res = Self { stack };
+            // push caller information
+            res.push();
+            res
+        } else {
+            Self {
+                stack: thin_vec![StackedErrorItem::new(e, Some(Location::caller()))],
+            }
         }
     }
 
-    pub fn from_err_locationless<E: Display + Send + Sync + 'static>(e: E) -> Self {
-        Self {
-            stack: thin_vec![ErrorItem::new(e, None)],
+    pub fn from_err_locationless<E: Display + Send + Sync + 'static>(mut e: E) -> Self {
+        if let Some(stack) = take_error_stack(&mut e) {
+            Self { stack }
+        } else {
+            Self {
+                stack: thin_vec![StackedErrorItem::new(e, None)],
+            }
         }
     }
 
@@ -198,8 +227,14 @@ impl Error {
 
     /// Pushes error `e` with location to the stack
     #[track_caller]
-    pub fn push_err<E: Display + Send + Sync + 'static>(&mut self, e: E) {
-        self.stack.push(ErrorItem::new(e, Some(Location::caller())));
+    pub fn push_err<E: Display + Send + Sync + 'static>(&mut self, mut e: E) {
+        if let Some(mut stack) = take_error_stack(&mut e) {
+            self.stack.append(&mut stack);
+            self.push();
+        } else {
+            self.stack
+                .push(StackedErrorItem::new(e, Some(Location::caller())));
+        }
     }
 
     /// Adds error `e` with location to the stack
@@ -210,8 +245,12 @@ impl Error {
     }
 
     /// Pushes error `e` without location information to the stack
-    pub fn push_err_locationless<E: Display + Send + Sync + 'static>(&mut self, e: E) {
-        self.stack.push(ErrorItem::new(e, None));
+    pub fn push_err_locationless<E: Display + Send + Sync + 'static>(&mut self, mut e: E) {
+        if let Some(mut stack) = take_error_stack(&mut e) {
+            self.stack.append(&mut stack);
+        } else {
+            self.stack.push(StackedErrorItem::new(e, None));
+        }
     }
 
     /// Adds error `e` without location information to the stack
@@ -259,19 +298,19 @@ impl Error {
     }
 
     /// Iteration over the [StackedErrorDowncast] items of `self`
-    pub fn iter(&self) -> Iter<'_, ErrorItem> {
+    pub fn iter(&self) -> Iter<'_, StackedErrorItem> {
         self.stack.iter()
     }
 
     /// Mutable iteration over the [StackedErrorDowncast] items of `self`
-    pub fn iter_mut(&mut self) -> IterMut<'_, ErrorItem> {
+    pub fn iter_mut(&mut self) -> IterMut<'_, StackedErrorItem> {
         self.stack.iter_mut()
     }
 }
 
 impl<'a> IntoIterator for &'a Error {
-    type IntoIter = Iter<'a, ErrorItem>;
-    type Item = &'a ErrorItem;
+    type IntoIter = Iter<'a, StackedErrorItem>;
+    type Item = &'a StackedErrorItem;
 
     fn into_iter(self) -> Self::IntoIter {
         self.iter()
@@ -279,8 +318,8 @@ impl<'a> IntoIterator for &'a Error {
 }
 
 impl<'a> IntoIterator for &'a mut Error {
-    type IntoIter = IterMut<'a, ErrorItem>;
-    type Item = &'a mut ErrorItem;
+    type IntoIter = IterMut<'a, StackedErrorItem>;
+    type Item = &'a mut StackedErrorItem;
 
     fn into_iter(self) -> Self::IntoIter {
         self.iter_mut()
@@ -288,9 +327,9 @@ impl<'a> IntoIterator for &'a mut Error {
 }
 
 impl Default for Error {
-    #[track_caller]
+    /// Uses `Error::empty()`
     fn default() -> Self {
-        Error::new()
+        Error::empty()
     }
 }