+++ src/macros.rs
@@ -1,32 +1,43 @@
 /// Equivalent to `return Err(Error::from_err(format_args!(...)))` if a string
 /// literal, `return Err(Error::from_err(expr))` if a single expression, or
 /// `return Err(Error::from_err(format!(...)))` otherwise.
+///
+/// If called with no arguments, "explicit bail"` is used as the message,
+/// analogous to how `panic!()` uses "explicit panic". If only the location
+/// matters, use `bail!(UnitError {})` instead, which formats to just the
+/// location.
 #[macro_export]
 macro_rules! bail {
+    () => {
+        return $crate::__private::Err($crate::Error::from_err("explicit bail"))
+    };
     ($msg:literal $(,)?) => {
-        return Err($crate::__private::format_err($crate::__private::format_args!($msg)))
+        return $crate::__private::Err($crate::__private::format_err($crate::__private::format_args!($msg)))
     };
     ($err:expr $(,)?) => {
-        return Err($crate::Error::from_err($err))
+        return $crate::__private::Err($crate::Error::from_err($err))
     };
     ($fmt:expr, $($arg:tt)*) => {
-        return Err($crate::Error::from_err($crate::__private::format!($fmt, $($arg)*)))
+        return $crate::__private::Err($crate::Error::from_err($crate::__private::format!($fmt, $($arg)*)))
     };
 }
 
 /// The `bail` macro but with `_locationless` variations
 #[macro_export]
 macro_rules! bail_locationless {
+    () => {
+        return $crate::__private::Err($crate::Error::from_err_locationless("explicit bail"))
+    };
     ($msg:literal $(,)?) => {
-        return Err($crate::__private::format_err_locationless(
+        return $crate::__private::Err($crate::__private::format_err_locationless(
             $crate::__private::format_args!($msg)
         ))
     };
     ($err:expr $(,)?) => {
-        return Err($crate::Error::from_err_locationless($err))
+        return $crate::__private::Err($crate::Error::from_err_locationless($err))
     };
     ($fmt:expr, $($arg:tt)*) => {
-        return Err($crate::Error::from_err_locationless(
+        return $crate::__private::Err($crate::Error::from_err_locationless(
             $crate::__private::format!($fmt, $($arg)*)
         ))
     };
@@ -65,18 +76,18 @@ macro_rules! anyhow {
 /// stackable error otherwise.
 ///
 /// Has `return Err(...)` with a [stacked_errors::Error](crate::Error) and
-/// attached location if the expression is false. An custom message can be
-/// attached that is used as a [StackableErr](crate::StackableErr) argument.
+/// attached location if the expression is false. A custom message can be
+/// attached, which takes all the same forms that [bail](crate::bail) does.
 ///
 /// ```
-/// use stacked_errors::{ensure, Result, StackableErr};
+/// use stacked_errors::{ensure, Result};
 ///
 /// fn ex(val0: bool, val1: bool) -> Result<()> {
 ///     ensure!(true);
 ///
 ///     ensure!(val0);
 ///
-///     ensure!(val1, format!("val1 was {}", val1));
+///     ensure!(val1, "val1 was {val1}");
 ///
 ///     Ok(())
 /// }
@@ -97,30 +108,29 @@ macro_rules! anyhow {
 /// ```
 #[macro_export]
 macro_rules! ensure {
-    ($expr:expr) => {
+    ($expr:expr $(,)?) => {
         if !$expr {
-            return Err($crate::Error::from_err($crate::__private::concat!(
+            return $crate::__private::Err($crate::Error::from_err($crate::__private::concat!(
                 "ensure(",
                 $crate::__private::stringify!($expr),
                 ") -> assertion failed"
             )));
         }
     };
-    ($expr:expr, $msg:expr) => {
+    ($expr:expr, $($arg:tt)+) => {
         if !$expr {
-            return Err($crate::Error::from_err($msg));
+            $crate::bail!($($arg)+);
         }
     };
 }
 
 /// Asserts that two expressions are equal to each other (with [PartialEq]),
-/// returning a stackable error if they are equal. [Debug] is also required if
-/// there is no custom message.
+/// returning a stackable error if they are not equal. [Debug] is also required
+/// if there is no custom message.
 ///
 /// Has `return Err(...)` with a [stacked_errors::Error](crate::Error) and
 /// attached location if the expressions are unequal. A custom message can be
-/// attached that is used as an [Error::from_err](crate::Error::from_err)
-/// argument.
+/// attached, which takes all the same forms that [bail](crate::bail) does.
 ///
 /// ```
 /// use stacked_errors::{ensure_eq, Result, StackableErr};
@@ -154,26 +164,28 @@ macro_rules! ensure {
 /// ```
 #[macro_export]
 macro_rules! ensure_eq {
-    ($lhs:expr, $rhs:expr) => {
+    ($lhs:expr, $rhs:expr $(,)?) => {
         // use the strategy that the core library does for `assert_eq`
         match (&$lhs, &$rhs) {
             (lhs, rhs) => {
                 // use the double inversion because we are relying on `PartialEq`
                 if !(*lhs == *rhs) {
-                    return Err($crate::Error::from_err($crate::__private::format!(
-                        "ensure_eq(\n lhs: {:?}\n rhs: {:?}\n) -> equality assertion failed",
-                        lhs,
-                        rhs,
-                    )));
+                    return $crate::__private::Err($crate::Error::from_err(
+                        $crate::__private::format!(
+                            "ensure_eq(\n lhs: {:?}\n rhs: {:?}\n) -> equality assertion failed",
+                            lhs,
+                            rhs,
+                        ),
+                    ));
                 }
             }
         }
     };
-    ($lhs:expr, $rhs:expr, $msg:expr) => {
+    ($lhs:expr, $rhs:expr, $($arg:tt)+) => {
         match (&$lhs, &$rhs) {
             (lhs, rhs) => {
                 if !(*lhs == *rhs) {
-                    return Err($crate::Error::from_err($msg));
+                    $crate::bail!($($arg)+);
                 }
             }
         }
@@ -186,8 +198,7 @@ macro_rules! ensure_eq {
 ///
 /// Has `return Err(...)` with a [stacked_errors::Error](crate::Error) and
 /// attached location if the expressions are equal. A custom message can be
-/// attached that is used as an [Error::from_err](crate::Error::from_err)
-/// argument.
+/// attached, which takes all the same forms that [bail](crate::bail) does.
 ///
 /// ```
 /// use stacked_errors::{ensure_ne, Result, StackableErr};
@@ -221,26 +232,28 @@ macro_rules! ensure_eq {
 /// ```
 #[macro_export]
 macro_rules! ensure_ne {
-    ($lhs:expr, $rhs:expr) => {
+    ($lhs:expr, $rhs:expr $(,)?) => {
         // use the strategy that the core library does for `assert_ne`
         match (&$lhs, &$rhs) {
             (lhs, rhs) => {
                 // use the double inversion because we are relying on `PartialEq`
                 if !(*lhs != *rhs) {
-                    return Err($crate::Error::from_err($crate::__private::format!(
-                        "ensure_ne(\n lhs: {:?}\n rhs: {:?}\n) -> inequality assertion failed",
-                        lhs,
-                        rhs,
-                    )));
+                    return $crate::__private::Err($crate::Error::from_err(
+                        $crate::__private::format!(
+                            "ensure_ne(\n lhs: {:?}\n rhs: {:?}\n) -> inequality assertion failed",
+                            lhs,
+                            rhs,
+                        ),
+                    ));
                 }
             }
         }
     };
-    ($lhs:expr, $rhs:expr, $msg:expr) => {
+    ($lhs:expr, $rhs:expr, $($arg:tt)+) => {
         match (&$lhs, &$rhs) {
             (lhs, rhs) => {
                 if !(*lhs != *rhs) {
-                    return Err($crate::Error::from_err($msg));
+                    $crate::bail!($($arg)+);
                 }
             }
         }
@@ -249,6 +262,9 @@ macro_rules! ensure_ne {
 
 /// Applies `get` and `stack_err_with(...)?` in a chain, this is compatible with
 /// many things.
+///
+/// The target can be any expression if it is parenthesized, e.x.
+/// `stacked_get!((self.value)["x"])`.
 ///
 /// ```
 /// use serde_json::Value;
@@ -278,6 +294,12 @@ macro_rules! ensure_ne {
 ///     ensure!(stacked_get!(value["State"]["Status"]) == "running");
 ///     ensure!(stacked_get!(value["State"]["Running"]) == true);
 ///
+///     struct Holder {
+///         value: Value,
+///     }
+///     let holder = Holder { value };
+///     ensure!(stacked_get!((holder.value)["State"]["Status"]) == "running");
+///
 ///     Ok(())
 /// }
 ///
@@ -295,7 +317,7 @@ macro_rules! ensure_ne {
 /// ```
 #[macro_export]
 macro_rules! stacked_get {
-    ($value:ident [$inx0:expr] $([$inx1:expr])*) => {{
+    ($value:tt [$inx0:expr] $([$inx1:expr])*) => {{
         // this is unrolled once to avoid a let binding
         // and allow multiple kinds of borrowing
         #[allow(unused)]
@@ -321,6 +343,9 @@ macro_rules! stacked_get {
 
 /// Applies `get_mut` and `stack_err_with(...)?` in a chain, this is compatible
 /// with many things.
+///
+/// The target can be any expression if it is parenthesized, e.x.
+/// `stacked_get_mut!((self.value)["x"])`.
 ///
 /// ```
 /// use serde_json::Value;
@@ -375,7 +400,7 @@ macro_rules! stacked_get {
 /// ```
 #[macro_export]
 macro_rules! stacked_get_mut {
-    ($value:ident [$inx0:expr] $([$inx1:expr])*) => {{
+    ($value:tt [$inx0:expr] $([$inx1:expr])*) => {{
         #[allow(unused)]
         let mut tmp = $crate::StackableErr::stack_err_with($value.get_mut($inx0),
             || $crate::__private::format!(